Further implementations and changes in relation to Gateway
Changes:

* Implemented getWorkers in Gateway
* Replaced all smart pointers in Gateway src and header with the raw data types
* Changed data types in Specs struct header

+++ include/database/Specs.h
@@ -6,13 +6,13 @@ namespace balancedbanana {
 		//Specifications (hardware) of a worker
 		struct Specs {
 			//Drive space
-			int space;
+			u_int64_t space;
 
 			//Random Access Memory
-			int ram;
+			uint64_t ram;
 
 			//CPU cores
-			int cores;
+			uint64_t cores;
 
 		};
 	}
